组装单 支付信息 ok 后台已自测

+++ ses-app/ses-web-ros/src/main/java/com/redescooter/ses/web/ros/dm/OpeAssemblyOrder.java
@@ -6,13 +6,14 @@ import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
-import java.io.Serializable;
-import java.math.BigDecimal;
-import java.util.Date;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.io.Serializable;
+import java.math.BigDecimal;
+import java.util.Date;
 
 @ApiModel(value = "com-redescooter-ses-web-ros-dm-OpeAssemblyOrder")
 @Data
@@ -83,7 +84,7 @@ public class OpeAssemblyOrder implements Serializable {
      */
     @TableField(value = "processing_fee_ratio")
     @ApiModelProperty(value = "加工费比例")
-    private Integer processingFeeRatio;
+    private BigDecimal processingFeeRatio;
 
     /**
      * 加工费
